move tx action key methods to transaction util
